feat(tabs): introduce tabs category and move fluid-tabs
Adds a dedicated Tabs docs category, migrates fluid-tabs out of the
card group, and updates landing imports and links to the new path.
