+++ content/docs/tabs/fluid-tabs.mdx
@@ -4,7 +4,7 @@ description: The component is a sliding animation card
 author: RudraSankha
 ---
 
-<ComponentPreview name="card-fluid-tabs--primary" />
+<ComponentPreview name="tabs-fluid-tabs--primary" />
 
 ## Installation
 
@@ -17,17 +17,17 @@ npm install framer-motion lucide-react
 
 <Step>Run the following command</Step>
 
-It will create a new file `fluid-tabs.tsx` inside the `components/animata/card` directory.
+It will create a new file `fluid-tabs.tsx` inside the `components/animata/tabs` directory.
 
 ```bash
-mkdir -p components/animata/card && touch components/animata/card/fluid-tabs.tsx
+mkdir -p components/animata/tabs && touch components/animata/tabs/fluid-tabs.tsx
 ```
 
 <Step>Paste the code</Step>{" "}
 
 Open the newly created file and paste the following code:
 
-```jsx file=<rootDir>/animata/card/fluid-tabs.tsx
+```jsx file=<rootDir>/animata/tabs/fluid-tabs.tsx
 
 ```
